fix(overlay): budget counters as int64, not float (exact OVB-AC-5 sum)
Review (CodeRabbit, Major): the budget counters were `number` → generated
float32, which rounds independently above 2^24 and could make the per-source
breakdown fail to sum to `consumed` on the wire (violating OVB-AC-5). Declare
every counter — REST consumed/limit, the three source counts, and Search
consumed/limit — as `integer, format: int64`; they are exact integer Redis
counts / int caps. Fix the whole family, including the pre-existing
consumed/limit (were float32), not just the new fields. budgetToWire now uses
exact int64 conversions (the domain values are already int — no float step).
Regenerated api_gen.go + FE schema.d.ts. Response narrowing (number→integer) is
non-breaking (contract-breaking-check passes).

Also (Minor): the OVB-AC-1 contract test now asserts Search.Window == "1s".

+++ backend/internal/contracts/api_gen.go
@@ -9791,20 +9791,20 @@ type OrganizationProfileFieldListResponse struct {
 // OverlayBudget The incumbent REST budget window's consumption and degradation band, its per-source breakdown, honest headroom, and the per-second Search window (overlay-budget.md "The budget read (wire shape)", OVB-AC-1/AC-5).
 type OverlayBudget struct {
 	Band     *OverlayBudgetBand `json:"band,omitempty"`
-	Consumed *float32           `json:"consumed,omitempty"`
+	Consumed *int64             `json:"consumed,omitempty"`
 
 	// Headroom Free REST capacity derived from our own counts, or the `~unknown` sentinel (OVB-PARAM-5) when a share cannot be attributed — never a fabricated number (OVB-AC-1).
-	Headroom *string  `json:"headroom,omitempty"`
-	Limit    *float32 `json:"limit,omitempty"`
+	Headroom *string `json:"headroom,omitempty"`
+	Limit    *int64  `json:"limit,omitempty"`
 
 	// Search The per-second Search-API window — metered, not gated, in branch 1, so the admin surface sees search pressure alongside REST.
 	Search *OverlayBudgetSearch `json:"search,omitempty"`
 
-	// Sources Per-source REST breakdown; the values sum exactly to `consumed` (OVB-AC-5). Absent sources have spent nothing this window.
+	// Sources Per-source REST breakdown; the values sum exactly to `consumed` (OVB-AC-5). Absent sources have spent nothing this window. Integer counts, so the per-source sum equals `consumed` exactly (a float would round independently and could break the sum above 2^24).
 	Sources *struct {
-		Capture    *float32 `json:"capture,omitempty"`
-		ForceFresh *float32 `json:"force_fresh,omitempty"`
-		Poller     *float32 `json:"poller,omitempty"`
+		Capture    *int64 `json:"capture,omitempty"`
+		ForceFresh *int64 `json:"force_fresh,omitempty"`
+		Poller     *int64 `json:"poller,omitempty"`
 	} `json:"sources,omitempty"`
 	Window *string `json:"window,omitempty"`
 }
@@ -9815,8 +9815,8 @@ type OverlayBudgetBand string
 // OverlayBudgetSearch The per-second Search-API window — metered, not gated, in branch 1, so the admin surface sees search pressure alongside REST.
 type OverlayBudgetSearch struct {
 	Band     *OverlayBudgetSearchBand `json:"band,omitempty"`
-	Consumed *float32                 `json:"consumed,omitempty"`
-	Limit    *float32                 `json:"limit,omitempty"`
+	Consumed *int64                   `json:"consumed,omitempty"`
+	Limit    *int64                   `json:"limit,omitempty"`
 	Window   *string                  `json:"window,omitempty"`
 }
